Added `ggrs_request_handlers` to handle requests for all session types
The handling of GGRS requests was the exact same among each of the session types. So from now on this one module will handle these requests.
